Move NetDev utils from common/ to net/
